add totient permutation problem
Although it is easy to come to the idea that the solution is the product
of two primes, traversing the list of primes in a double for-loop is just
as expensive as calculating the totitent function values. Reducing the
range of primes is feasible, but did not find any reasonable explanation,
so instead of using any 'educated guess', I rather post this solution, since
this is for sure correct.

Signed-off-by: Tamas Kurics <tamas.kurics@gmail.com>